пофиксил инсталл точка Ш
